Revert "PMM-9803 incorrect connection message (#390)"
This reverts commit 04325334753c447c88c7a194feec8358a8b30233.
