+++ public/app/percona/shared/components/Elements/PlatformConnectedLoader/PlatformConnectedLoader.constants.ts
