+++ public/app/percona/shared/components/Elements/PlatformConnectedLoader/PlatformConnectedLoader.messages.ts
@@ -1,3 +1,4 @@
 export const Messages = {
-  platformUser: 'Connect with Percona Account to access this content',
+  notConnected: 'Not connected to Portal. You can connect in',
+  portalSettings: 'Platform Settings.',
 };
