+++ public/app/percona/shared/components/Elements/PlatformConnectedLoader/PlatformConnectedLoader.styles.ts
