--- public/app/percona/shared/components/Elements/PlatformConnectedLoader/components/ErrorMessage/ErrorMessage.messages.ts
@@ -1,4 +1,0 @@
-export const Messages = {
-  notConnected: 'Not connected to Portal. You can connect in',
-  portalSettings: 'Platform Settings.',
-};
